+++ QuestionsServer/QuestionsData/DTO/QuestionDto.cs
@@ -6,13 +6,13 @@ namespace QuestionsData.DTO
 {
   public class QuestionDto
   {
-    public int QuestionId { get; set; }
+    public long QuestionId { get; set; }
     public string Condition { get; set; }
     public List<OptionDto> Options { get; set; }
     public string OriginalBackgroundUrl { get; set; }
     public QuestionType QuestionType { get; set; }
     public DateTime QuestionAddDate { get; set; }
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public string Login { get; set; }
     public SexType SexType { get; set; }
     public string SmallAvatarUrl { get; set; }
