+++ QuestionsServer/QuestionsData/Entities/OptionEntity.cs
@@ -9,7 +9,7 @@ namespace QuestionsData.Entities
     [Key]
     public int OptionId { get; set; }
 
-    public int QuestionId { get; set; }
+    public long QuestionId { get; set; }
     public string Text { get; set; }
     public int Position { get; set; }
     public int AnswersCount { get; set; }
