+++ QuestionsServer/QuestionsData/Repositories/CommentsRepository.cs
@@ -20,7 +20,7 @@ namespace QuestionsData.Repositories
       _db = db;
     }
 
-    public async Task<int> AddComment(int userId, int questionId, string comment, int? previousCommentId)
+    public async Task<int> AddComment(long userId, long questionId, string comment, int? previousCommentId)
     {
       var commentedTime = DateTime.UtcNow;
 
@@ -31,12 +31,12 @@ namespace QuestionsData.Repositories
       return (int) commentedId.Value;
     }
 
-    public async Task<bool> UpdateCommentFeedback(int userId, int commentId, QuestionFeedbackType feedback)
+    public async Task<bool> UpdateCommentFeedback(long userId, int commentId, QuestionFeedbackType feedback)
     {
       return await _db.Database.ExecuteSqlCommandAsync($"updateCommentFeedback {userId}, {commentId}, {feedback}") > 0;
     }
 
-    public async Task<List<CommentDb>> GetComments(int userId, int questionId, int offset, int count)
+    public async Task<List<CommentDb>> GetComments(long userId, long questionId, int offset, int count)
     {
       return await _db.CommentDb.AsNoTracking()
                .FromSql($"select * from dbo.getComments({userId}, {questionId})").OrderByDescending(x=>x.CommentedDate).Skip(offset).Take(count).ToListAsync() ??
